Remove old DialogBase as everything is in modern format now
