+++ src/lib-editor/include/Interfaces/DialogInterfaces.hpp
@@ -1,6 +1,5 @@
 #pragma once
 
-#include "Dialogs/DialogBase.hpp"
 #include <Memory.hpp>
 #include <filesystem>
 #include <functional>
@@ -31,26 +30,4 @@ public:
 
 public:
     virtual void open(const std::string& text) = 0;
-};
-
-class PlaytestSettingsDialogInterface : public DialogInterface
-{
-public:
-    PlaytestSettingsDialogInterface(
-        mem::Rc<Gui> gui,
-        const std::string& id,
-        const std::string& title,
-        const std::vector<OptionLine>& options)
-        : DialogInterface(gui, id, title, options)
-    {
-    }
-
-    virtual ~PlaytestSettingsDialogInterface() = default;
-
-public:
-    [[nodiscard]] virtual std::filesystem::path getBinaryPath() const = 0;
-
-    [[nodiscard]] virtual std::string getLaunchParameters() const = 0;
-
-    [[nodiscard]] virtual std::filesystem::path getWorkingDirPath() const = 0;
 };
