+++ src/lib-editor/src/Dialogs/EditPropertyDialog.cpp
@@ -5,7 +5,7 @@ static inline constexpr const char* EDIT_DIALOG_ID = "EditPropertyDialog";
 
 EditPropertyDialog::EditPropertyDialog(
     mem::Rc<Gui> gui, mem::Box<ToolPropertyInterface> property)
-    : ModernDialogInterface(
+    : DialogInterface(
         gui, EDIT_DIALOG_ID, Strings::Editor::EDIT_PROPERTIES)
     , property(std::move(property))
 {
